Include error cause in exception if available
Low level errors, e.g. DNS lookup failure are reported by httpkit-client in the
`:error` key of the response.

This patch adds this error as the cause of the generated exception to allow for
reporting and detection of the reason of failure.

Previously in some cases there was no detail on why the request has failed.

Signed-off-by: Philipp Meier <philipp.meier@nubank.com.br>